# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Tiny Ledger
  description: A simple ledger API
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'

paths:
  /v1/accounts:
    post:
      summary: Open an account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountOpeningRequest'
      responses:
        '201':
          description: Account open accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'

  /v1/accounts/{accountId}/balances:
    get:
      summary: Display account balance
      parameters:
        - $ref: '#/components/parameters/AccountIdPathParameter'
      responses:
        '200':
          description: Display account balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/accounts/{accountId}/transactions:
    get:
      summary: Display account transactions
      parameters:
        - $ref: '#/components/parameters/AccountIdPathParameter'
        - $ref: '#/components/parameters/PageNumberQueryParameter'
        - $ref: '#/components/parameters/PageSizeQueryParameter'
        - $ref: '#/components/parameters/PageSortParameter'
      responses:
        '200':
          description: Display account transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Move money from an account
      parameters:
        - $ref: '#/components/parameters/AccountIdPathParameter'
      requestBody:
        description: positive amounts for credit and negative for debit
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoneyMovement'
      responses:
        '204':
          description: Accepted
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Not enough balance

components:
  parameters:
    AccountIdPathParameter:
      name: accountId
      in: path
      description: account unique identifier
      required: true
      schema:
        $ref: '#/components/schemas/AccountId'
    PageNumberQueryParameter:
      name: page
      in: query
      description: page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQueryParameter:
      name: size
      in: query
      description: page size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PageSortParameter:
      in: query
      required: false
      name: sort
      description: Sort order
      schema:
        $ref: "#/components/schemas/PageSort"
        default: DESC
  schemas:
    AccountId:
      type: string
      format: uuid
      example: 5e00d80e-4b1f-4193-8edf-c7aab51f31c3
    PageSort:
      type: string
      enum: [ ASC, DESC ]
    Currency:
      type: string
      enum: [ EUR, USD, GBP, JPY ]

    AccountOpeningRequest:
      type: object
      required:
        - currency
      properties:
        currency:
            $ref: '#/components/schemas/Currency'

    AccountResponse:
      type: object
      required:
        - accountId
      properties:
        accountId:
          $ref: '#/components/schemas/AccountId'

    BalanceResponse:
      type: object
      required:
        - balance
      properties:
        balance:
          type: number
          format: double
          example: 100.11

    TransactionsResponse:
      type: object
      required:
        - transactions
        - page
        - totalPages
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        page:
          type: integer
          minimum: 1
        nextPage:
          type: integer
          minimum: 1
        totalPages:
          type: integer
          minimum: 1
          nullable: true


    Transaction:
      type: object
      required:
        - dateTime
        - amount
        - balance
      properties:
        dateTime:
          type: string
          format: date-time
          example: "2024-05-20T14:30:00+05:30"
        amount:
          type: number
          format: double
          example: 123.45
        balance:
          type: number
          format: double
          example: 123.45
        description:
          type: string
          example: "some reference to the transaction"

    MoneyMovement:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
          # Avoid zero-value transactions
          not:
            const: 0
        description:
          type: string
          maxLength: 255

    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "error message"